backend/pyproject.toml
[project]
name = "silicon-studio-backend"
version = "0.1.0"
description = "Local-first LLM fine-tuning backend for Perimeter.ai"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "mlx-lm>=0.1.0",
    "pandas>=2.2.0",
    "presidio-analyzer>=2.2.351",
    "presidio-anonymizer>=2.2.351",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.2.0",
    "psutil>=5.9.8",
    "mcp>=1.0.0",
    "onnxruntime>=1.17.0",
    "tokenizers>=0.19.0",
    "rank-bm25>=0.2.2",
    "numpy>=1.26.0",
    "aiohttp>=3.9.0",
    "trafilatura>=1.8.0",
    "duckduckgo-search>=6.0.0",
    "aiofiles>=23.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "black>=24.0.0",
    "isort>=5.13.0"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools]
packages = ["app"]
